chore: Move the utilities collections to their own directory
