+++ Assets/Scripts/Utilities/Collections.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: da65c6fc152ad9146a91bfea677a6070
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
