+++ Assets/Scripts/Utilities/Collections/IIndexed.cs
@@ -1,0 +1,7 @@
+namespace Utilities.Collections
+{
+    public interface IIndexed
+    {
+        int Index { get; set; }
+    }
+}
