+++ Assets/Scripts/Utilities/Collections/NativePriorityQueue.cs
@@ -5,6 +5,7 @@ using Unity.Burst;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
 using Unity.Jobs;
+using Utilities.Collections;
 
 namespace Utilities
 {
