+++ Assets/Scripts/Utilities/Collections/NativePriorityQueueDisposeJob.cs
@@ -2,7 +2,7 @@ using System;
 using UnityEngine;
 using Unity.Jobs;
 
-namespace Utilities
+namespace Utilities.Collections
 {
     internal struct NativePriorityQueueDisposeJob<T> : IJob where T : unmanaged, IComparable<T>
     {
