--- Assets/Scripts/Utilities/Grid2DUtils.cs
@@ -1,9 +1,0 @@
-using NavigationGraph;
-
-public static class Grid2DUtils
-{
-    public static Cell GetCurrentCell()
-    {
-        return new Cell(); // Placeholder implementation
-    }
-}
