--- Assets/Scripts/Utilities/Grid2DUtils.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 22d516bdcf78e114eb792760c1337447
