--- Assets/Scripts/Utilities/IIndexed.cs
@@ -1,4 +1,0 @@
-public interface IIndexed
-{
-    int Index { get; set; }
-}
