--- Assets/Scripts/Utilities/NativeListExtensions.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 94bee2321aab45feb1252d30eebf0307
-timeCreated: 1736899406
