--- Assets/Scripts/Utilities/NativePriorityQueue.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: d127e5d449fe49dba1588793ad131ed0
-timeCreated: 1739390526
