--- Assets/Scripts/Utilities/SwapBack.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7b64f5c2f121bfd40a629caeaed467a9
