--- Assets/Scripts/Utilities/SwapBackList.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3f0e783b864b4864883c3f83fdcae0ed
