--- Assets/Scripts/Utilities/SwapBackListIndexed.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3f7bee3fb2c8de44c83fe0d72f0f0df3
